rewrite gateway CLass with autoload
